requirements.txt
amqp==5.2.0
asgiref==3.6.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2
django-allauth==0.50.0
django-otp==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==6.0.0
idna==3.4
inflection==0.5.1
kombu==5.3.7
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
prompt_toolkit==3.0.47
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.7
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
